set_dataset_processing_params to base classifier, removed it from resnet model.
