+++ src/super_gradients/training/models/classification_models/resnet.py
@@ -10,23 +10,15 @@ Pre-trained CIFAR10 models: 'deci-model-repository/CIFAR_NAS_#?_????_?/ckpt_best
 Code adapted from https://github.com/pytorch/vision/blob/master/torchvision/models/resnet.py
 """
 from collections import OrderedDict
-from functools import lru_cache
-from typing import Optional, List
 
 import torch.nn as nn
 import torch.nn.functional as F
 
-from super_gradients.common.decorators.factory_decorator import resolve_param
-from super_gradients.common.factories.processing_factory import ProcessingFactory
 from super_gradients.common.object_names import Models
 from super_gradients.common.registry.registry import register_model
 from super_gradients.modules.utils import width_multiplier
 from super_gradients.training.models.classification_models.base_classifer import BaseClassifier
-from super_gradients.training.pipelines.pipelines import ClassificationPipeline
-from super_gradients.training.processing.processing import Processing
 from super_gradients.training.utils import get_param
-from super_gradients.training.utils.media.image import ImageSource
-from super_gradients.training.utils.predict import ImagesPredictions
 from super_gradients.training.utils.regularization_utils import DropPath
 
 
@@ -238,63 +230,6 @@ class ResNet(BaseClassifier):
             self.linear = new_head
         else:
             self.linear = nn.Linear(width_multiplier(512, self.width_mult) * self.expansion, new_num_classes)
-
-    @resolve_param("image_processor", ProcessingFactory())
-    def set_dataset_processing_params(
-        self,
-        class_names: Optional[List[str]] = None,
-        image_processor: Optional[Processing] = None,
-    ) -> None:
-        """Set the processing parameters for the dataset.
-
-        :param class_names:     (Optional) Names of the dataset the model was trained on.
-        :param image_processor: (Optional) Image processing objects to reproduce the dataset preprocessing used for training.
-        :param iou:             (Optional) IoU threshold for the nms algorithm
-        :param conf:            (Optional) Below the confidence threshold, prediction are discarded
-        """
-        self._class_names = class_names or self._class_names
-        self._image_processor = image_processor or self._image_processor
-
-    @lru_cache(maxsize=1)
-    def _get_pipeline(self, fuse_model: bool = True) -> ClassificationPipeline:
-        """Instantiate the prediction pipeline of this model.
-        :param fuse_model: If True, create a copy of the model, and fuse some of its layers to increase performance. This increases memory usage.
-        """
-        if None in (self._class_names, self._image_processor):
-            raise RuntimeError(
-                "You must set the dataset processing parameters before calling predict.\n" "Please call `model.set_dataset_processing_params(...)` first."
-            )
-
-        pipeline = ClassificationPipeline(
-            model=self,
-            image_processor=self._image_processor,
-            class_names=self._class_names,
-            fuse_model=fuse_model,
-        )
-        return pipeline
-
-    def predict(self, images: ImageSource, fuse_model: bool = True) -> ImagesPredictions:
-        """Predict an image or a list of images.
-
-        :param images:  Images to predict.
-        :param iou:     (Optional) IoU threshold for the nms algorithm. If None, the default value associated to the training is used.
-        :param conf:    (Optional) Below the confidence threshold, prediction are discarded.
-                        If None, the default value associated to the training is used.
-        :param fuse_model: If True, create a copy of the model, and fuse some of its layers to increase performance. This increases memory usage.
-        """
-        pipeline = self._get_pipeline(fuse_model=fuse_model)
-        return pipeline(images)  # type: ignore
-
-    def predict_webcam(self, iou: Optional[float] = None, conf: Optional[float] = None, fuse_model: bool = True):
-        """Predict using webcam.
-
-        :param iou:     (Optional) IoU threshold for the nms algorithm. If None, the default value associated to the training is used.
-        :param conf:    (Optional) Below the confidence threshold, prediction are discarded.
-                        If None, the default value associated to the training is used.
-        :param fuse_model: If True, create a copy of the model, and fuse some of its layers to increase performance. This increases memory usage.
-        """
-        pipeline = self._get_pipeline(fuse_model=fuse_model)
-        pipeline.predict_webcam()
 
 
 @register_model(Models.RESNET18)
